+++ packages/cli/src/types/config-flags.ts
@@ -2,5 +2,4 @@ import { UserConfig } from './user-config'
 
 export interface ConfigFlags extends UserConfig {
     configName: string
-    noNamespace: boolean
 }