+++ packages/cli/templates/Gjs/ui/main.d.ts
@@ -1,9 +1,9 @@
 <% const Gio = dep.find('Gio') %>
 
-<%_ if(useNamespace){ _%>
-    import type Gio from '../<%= Gio.packageName %>.js';
-<%_ } else { _%>
+<%_ if(noNamespace){ _%>
     import type * as Gio from '../<%= Gio.packageName %>.js';
+<%_ } else { _%>
+    import type Gio from '../<%= Gio.packageName %>.js';
 <%_ } _%>
 
 import { ComponentManager } from './components/__init__.js';