+++ packages/cli/templates/Gjs/ui/search.d.ts
@@ -3,12 +3,12 @@
 <% const St = dep.find('St') %>
 <% const Clutter = dep.find('Clutter') %>
 
-<%_ if(useNamespace){ _%>
-    import type St from '../<%= St.packageName %>.js';
-    import type Clutter from '../<%= Clutter.packageName %>.js';
-<%_ } else { _%>
+<%_ if(noNamespace){ _%>
     import type * as St from '../<%= St.packageName %>.js';
     import type * as Clutter from '../<%= Clutter.packageName %>.js';
+<%_ } else { _%>
+    import type St from '../<%= St.packageName %>.js';
+    import type Clutter from '../<%= Clutter.packageName %>.js';
 <%_ } _%>
 
 import { AppSearchProvider } from './appDisplay.js';